Move the info func to utils.bids.
